+++ src/main/java/dev/lambdaurora/lambdynlights/accessor/DynamicLightHandlerHolder.java
@@ -12,6 +12,7 @@ package dev.lambdaurora.lambdynlights.accessor;
 import dev.lambdaurora.lambdynlights.api.DynamicLightHandler;
 import dev.lambdaurora.lambdynlights.config.LightSourceSettingEntry;
 import net.minecraft.network.chat.Text;
+import net.minecraft.resources.Identifier;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.level.block.entity.BlockEntity;
@@ -29,6 +30,8 @@ public interface DynamicLightHandlerHolder<T> {
 	LightSourceSettingEntry lambdynlights$getSetting();
 
 	Text lambdynlights$getName();
+
+	Identifier lambdynlights$getId();
 
 	@SuppressWarnings("unchecked")
 	static <T extends Entity> DynamicLightHandlerHolder<T> cast(EntityType<T> entityType) {