+++ src/main/java/dev/lambdaurora/lambdynlights/mixin/DevModeMixin.java
@@ -14,12 +14,14 @@ import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.renderer.GameRenderer;
+import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 
 @Mixin(GameRenderer.class)
 public class DevModeMixin {
+	@Final
 	@Shadow
 	private Minecraft minecraft;
 
@@ -32,8 +34,10 @@ public class DevModeMixin {
 		int bottom = this.minecraft.getWindow().getGuiScaledHeight();
 		int y = bottom - this.minecraft.font.lineHeight;
 
-		graphics.fill(0, y - 4, this.minecraft.font.width(text) + 4, bottom, 0xaa000000);
-		graphics.drawShadowedText(this.minecraft.font, text, 2, y - 2, 0xff0000);
+		if (this.minecraft.isGameLoadFinished()) {
+			graphics.fill(0, y - 4, this.minecraft.font.width(text) + 4, bottom, 0xaa000000);
+			graphics.drawShadowedText(this.minecraft.font, text, 2, y - 2, 0xff0000);
+		}
 		original.call(graphics);
 	}
 }